requirements.txt
accelerate==0.30.1
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.7
cycler==0.12.1
datasets==2.18.0
Deprecated==1.2.15
dill==0.3.8
durationpy==0.9
einops==0.8.0
et_xmlfile==2.0.0
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.5.4
gdown==5.2.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.5
kubernetes==31.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
mdurl==0.1.2
mmh3==5.0.1
modelscope==1.14.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.20.1
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.11
oss2==2.18.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.2.2
posthog==3.7.2
protobuf==5.28.3
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.0.1
torch==2.5.1
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.67.0
transformers==4.46.3
transformers_stream_generator
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==14.1
wikipedia==1.4.0
Wikipedia-API==0.7.1
wrapt==1.16.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.21.0

# Commented duplicates:
#beautifulsoup4==4.12.3
#certifi==2024.2.2
#charset-normalizer==3.3.2
#filelock==3.14.0
#fsspec==2024.2.0
#huggingface-hub==0.23.2
#idna==3.7
#importlib_metadata==7.1.0
#Jinja2==3.1.4
#joblib==1.4.2
#MarkupSafe==2.1.5
#mpmath==1.3.0
#numpy==1.26.4
#packaging==24.0
#pandas==2.2.2
#pillow==10.3.0
#python-dateutil==2.9.0.post0
#pytz==2024.1
#PyYAML==6.0.1
#regex==2024.5.15
#requests==2.32.2
#safetensors==0.4.3
#scikit-learn==1.5.0
#scipy==1.13.1
#setuptools==70.0.0
#six==1.16.0
#soupsieve==2.5
#sympy==1.12
#threadpoolctl==3.5.0
#tokenizers==0.19.1
#torch==2.3.0
#tqdm==4.66.4
#transformers==4.41.1
#typing_extensions==4.12.0
#tzdata==2024.1
#urllib3==2.2.1
#zipp==3.19.0